Split :feature:detail module to api and impl
