+++ feature/detail/api/.gitignore
