+++ feature/detail/api/build.gradle
@@ -1,0 +1,19 @@
+plugins {
+    id "moop.android.library"
+    id "moop.android.compose"
+    id "moop.android.hilt"
+}
+
+android {
+    namespace "soup.movie.feature.detail"
+}
+
+dependencies {
+    implementation projects.core.kotlin
+    implementation projects.core.designsystem
+    implementation projects.core.resources
+
+    implementation libs.kotlin.stdlib
+
+    implementation libs.compose.foundation
+}
