+++ feature/detail/api/src/main/AndroidManifest.xml
