+++ feature/detail/api/src/main/java/soup/movie/feature/detail/DetailComposableFactory.kt
@@ -1,0 +1,45 @@
+/*
+ * Copyright 2024 SOUP
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package soup.movie.feature.detail
+
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.remember
+import androidx.compose.ui.platform.LocalContext
+import dagger.hilt.EntryPoint
+import dagger.hilt.InstallIn
+import dagger.hilt.android.EntryPointAccessors
+import dagger.hilt.components.SingletonComponent
+
+interface DetailComposableFactory {
+    @Composable
+    fun DetailNavGraph()
+}
+
+@Composable
+fun rememberDetailComposableFactory(): DetailComposableFactory {
+    val context = LocalContext.current
+    return remember(context) {
+        EntryPointAccessors
+            .fromApplication(context, DetailComposableFactoryEntryPoint::class.java)
+            .detailComposableFactory()
+    }
+}
+
+@EntryPoint
+@InstallIn(SingletonComponent::class)
+interface DetailComposableFactoryEntryPoint {
+    fun detailComposableFactory(): DetailComposableFactory
+}
