+++ feature/detail/impl/src/main/java/soup/movie/feature/detail/impl/DetailComposableFactoryImpl.kt
@@ -1,0 +1,31 @@
+/*
+ * Copyright 2024 SOUP
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package soup.movie.feature.detail.impl
+
+import androidx.compose.runtime.Composable
+import androidx.hilt.navigation.compose.hiltViewModel
+import soup.movie.feature.detail.DetailComposableFactory
+import javax.inject.Inject
+
+class DetailComposableFactoryImpl @Inject constructor() : DetailComposableFactory {
+
+    @Composable
+    override fun DetailNavGraph() {
+        DetailNavGraph(
+            viewModel = hiltViewModel(),
+        )
+    }
+}
