+++ feature/detail/impl/src/main/java/soup/movie/feature/detail/impl/DetailNavGraph.kt
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package soup.movie.feature.detail
+package soup.movie.feature.detail.impl
 
 import androidx.compose.animation.AnimatedVisibility
 import androidx.compose.animation.fadeIn
