+++ feature/detail/impl/src/main/java/soup/movie/feature/detail/impl/di/FeatureDetailModule.kt
@@ -1,0 +1,33 @@
+/*
+ * Copyright 2024 SOUP
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package soup.movie.feature.detail.impl.di
+
+import dagger.Binds
+import dagger.Module
+import dagger.hilt.InstallIn
+import dagger.hilt.components.SingletonComponent
+import soup.movie.feature.detail.DetailComposableFactory
+import soup.movie.feature.detail.impl.DetailComposableFactoryImpl
+
+@Module
+@InstallIn(SingletonComponent::class)
+interface FeatureDetailModule {
+
+    @Binds
+    fun bindsDetailComposableFactoryImpl(
+        impl: DetailComposableFactoryImpl,
+    ): DetailComposableFactory
+}
